requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
billiard==4.2.0
boto3==1.34.162
botocore==1.34.162
cachetools==5.5.0
celery==5.4.0
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==43.0.3
distro==1.9.0
Django==5.1
django-allow-cidr==0.7.1
django-celery-results==2.5.1
django-cors-headers==4.4.0
django-environ==0.11.2
django-non-dark-admin==2.0.2
django-otp==1.5.4
django-ses==4.1.0
django-simple-history==3.7.0
django-storages==1.14.4
djangorestframework==3.15.2
dnspython==2.7.0
docusign-esign==4.0.0
email_validator==2.2.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.1
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.24.6
idna==3.8
importlib_metadata==8.4.0
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.4.0
litellm==1.43.9
MarkupSafe==2.1.5
multidict==6.0.5
nose==1.3.7
numpy==2.1.2
oauthlib==3.2.2
openai==1.42.0
opencv-python==4.10.0.84
openpyxl==3.1.5
packaging==24.1
pillow==10.4.0
prompt_toolkit==3.0.47
proto-plus==1.25.0
protobuf==5.29.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.9.0
pyparsing==3.2.0
pypdf==4.3.1
pypdfium2==4.30.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
rapidfuzz==3.9.6
redis==5.0.8
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.20.0
rsa==4.9
ruff==0.3.4
s3transfer==0.10.2
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
tenacity==9.0.0
thefuzz==0.22.1
tiktoken==0.7.0
tokenizers==0.20.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
yarl==1.9.4
zipp==3.20.0
